Set encoding to utf-8
Vim assumes files are latin-1 unless specified otherwise. UTF-8 is the standard going forward and works better with all languages. Encoding is how vim will represent characters internally, aka what will be show to you in the terminal.

Similar is `fileencoding` which sets the encoding of the file in the buffer.