Add vimrc autogroup and go to last edited place in file
When editing a file, always jump to the last known cursor position.
Don't do it for commit messages, when the position is invalid, or when
inside an event handler (happens when dropping a file on gvim).

augroup gives you a place to collect autocmds to avoid duplication and provide organization.
autocmd! clears previous setting of this group name.