Highlight leading and trailing spaces
Create three highlights for different kinds of spaces, the middle of which should match the background of the terminal so as to hide them. Turn on highlighting all spaces. Then use layered matching (match, 2match, 3match) to find the corresponding uses of spaces. Turn off the now unneeded trailing space character.