Handle whitespace to avoid matchparens
Matchparens uses 3match to set colors, which was causing a bunch of issues. The docs say to avoid 3match for this reason. Instead I've created a regex that looks for whitespace characters between non-whitespace characters and uses vim's zs/ze to handle what to send up for highlighting.

Confusingly, the docs suggest that the lower number takes precedence on match, however if I make WhiteSpaceBol the match2, it clobbers my other regex.

Also updated WhiteSpaceBol to highlight all whitespace characters, not just spaces.