Adjust wildmode
Wildmode, aka what you see when you tab complete in the vim commandline, is a comma separated list of options for what to display on hitting the wildchar. List: longest shows a horizontal list of possible completions. Full completes the next possible option.

In Neovim full respects the wildmenu setting, though as of 0.4.3 this is broken if it's the second in the list. Waiting on fixes that appear to be coming in 0.5.0.