Treat <p> like the block tag it is
Default indent settings are collected in $VIMRUNTIME/indent. (An env that only shows while vim is running). You can override/add to these settings using a variety of tags, some of which aren't listed in indent.txt in the help docs (try :h language-indent, e.g. html-indent) but that you'll find throughout people's vimrcs. The biggest culprit of this is `g:html_indent_tags`, which isn't the documented way to add or remove tags for indent (that goes to html_indent_inctags and html_indent_autotags). It's in TPope's ragtag as well as Thoughtbot's dotfiles. I'm going to guess it's either an older holdover or a mismatch.

You'll also notice some of these lists using '\|' instead of ',', which is another way of separating commmands.