Add minpac for upgraded package management
I originally had set vim-plug, but found minpac while researching. Minpac uses the built in package management and asynchrony of both vim and neovim, but includes the ease of use of something like vim-plug.

I'm also adding it here as a package so it can manage its own upgrades. It needs to be optional so packadd can do the loading in the first place.