Set the shell explicitly
Neovim, like Vim, relies on the SHELL env variable. My login shell is set to zsh, but SHELL refers to the process that spawned the login shell. The default shell, which on OSX at present is /bin/sh. I could modify the variable, but I'm not sure what else relies on it and though zsk is POSIX compliant I'd rather not worry about it.